requirements.txt
PyQt6
matplotlib
pandas
numpy 
python-binance